requirements.txt
Django==4.0.2
django-mptt==0.14.0
django-mptt-admin==2.3.0
